refac: move websocket/constructor/* to websocket/
