+++ websocket/rpc.go
@@ -1,4 +1,4 @@
-package constructor
+package websocket
 
 import (
 	"fmt"
